requirements.txt
gpt4all>=2.8.0
llama-cpp-python>=0.3.0
qdrant-client>=1.17.0
discord.py>=2.3.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
aiohttp>=3.11.0
pydantic>=2.0.0
requests>=2.31.0
PyYAML>=6.0.0
psutil>=5.9.0
sentence-transformers>=2.2.0
numpy>=1.24.0
